Moved user_passwords_requests.py from src/tests to src/tests/requests folder and fixed import in test_user_passwords.py.
