+++ src/tests/test_user_passwords.py
@@ -1,11 +1,9 @@
 import json
 
 from src.app import app
-from .requests.user_requests import UserRequests
 from .requests.basic_requests import BasicRequests
 from .requests.admin_requests import AdminRequests
-from .requests.login_requests import client, LoginRequests, resource
-from src.tests.user_passwords_requests import UserPasswords, PasswordResource
+from .requests.user_passwords_requests import UserPasswords, PasswordResource
 
 
 def test_register(client):
